+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.6.2] - 2021-08-05
+## Changes
+- added possibility reboot console
+- added possibility switch to Television input
+- code and config reconfigured
+
+## [1.6.0] - 2021-08-04
+## Changes
+- fixes
+
 ## [1.6.0] - 2021-08-04
 ## Changes
 - added possibility load inputs list direct from device